The following terms were curated:
    * optical_subsystem

definition_citation (should be _source): http://en.wikipedia.org/wiki/Optics Intro to Flow by BD www.petermac.org/pdf/Intro%20to%20Flow%20Cytometry%20.pdf
These are inappropriate (first case) or outdated (latter case) and were deleted and 'John Quinn' put there instead.

editor_note: DS: Is 'subsytem' necessary or is 'optical_system' enough.
Not sure its graph position since an optical subsystem is not necessarily an instrument, but more likely part of one.

    * diode_laser

Refined definition: A diode laser is a laser in which the active medium is a p-n junction semiconductor laser diode, similar to that found in a light-emitting diode. Laser diodes emit at wavelengths from 375 nm to 1800 nm, and wavelengths of over 3 micrometer have been demonstrated. A diode laser can by used to irradiate cells in a flow cytometer.
Updated def-editor

    * dye_laser

minor orthographic corrections in def.
Updated def-editor

    * gas_laser

Altered definition: A gas laser is a laser in which the lasing medium is a gas. The laser medium is places between two parallel mirrors for light emmission amplification. The gas is excited to emit light via an external light source or an electric current dischargeing through the gas.
Updated def-editor

    * ion_laser

Shortened definition (skipped empirical statement of class parenthood).
Updated def-editor

    * argon_ion_laser

normalized def
Updated def-editor

    * krypton_ion_laser

Normalized def
Updated def-editor

    * solid_state_laser

iMPROVED DEFINITION: A solid-state laser is a laser that uses a lasing medium that is a solid, rather than a liquid such as dye lasers or a gas such as gas lasers. Semiconductor-based diode lasers are also in the solid state, but are generally considered separately from solid-state lasers.
Updated def-editor

    * neodymium-YAG_laser

Updated def-editor

    * obscuration_bar

Updated def-editor

    * heating_block

New definition: A heating block is an instrument or part of an instrument which raises or maintains the temperature of a sample to a defined constant temperature during certain parts of an assay
Updated def-editor


Additionally I have curated:

    * laser

definition: A laser (acronym for light amplification by the stimulated emission of radiation) is a light source that emits photons in a coherent beam. A laser uses a solid, liquid or gaseous lasing medium, that contains molecules, of which some atoms have electrons that emit photons of the same coherent frequency when falling back to their normal orbital after excitation (stimulation) by external means. A laser is the most common way to irradiate a cell in a flow cytometer.


A general comment: I  don't think we should commit on a 'per change'-basis, but on a per task basis. Having 15 revisions for one task is not necessary I belive.