Changelog:
- material entity: 
editor note added to mark the submission to BFO -> term to be imported.

- molecular entities:
AI for BP: log a request to CHEBI to use singular 'molecular entity' -> term to be imported.

-chemical entities in solution/mixture:
AI for Alan to implement

- anatomical entity (formerly macroscopic part of multicellular organism):
now a defined class: union of (multi-tissue structure, body substance, portion of tissue_

Label changed to anatomical entity as it was felt that 'macroscopic part of multicellular organism' introduce an impractical delineation.

editor note added: OBI 'anatomical entity' would essentially correspond to FMA 'material anatomical entity' rather than FMA 'physical anatomical entity'. OBI does not include 'immaterial anatomical entity' as it would clash with the parent class 'material entity'
There is also a documentation note:
DOCUMENTATION NOTE: OBI anatomical entity granularity level excludes cell and biological  molecules.
 

- environmental matter:
left in place but branch core tag removed.

- entity of organismal origin
label changed to extracted entity of organismal origin during the call.
However I have the following questions:
entity of organismal origin:
Current definition= An entity of organismal origin is a biomaterial entity which is obtained from an organism.

alter definition to include the following:
An entity of organismal origin is a biomaterial entity which is obtained from an organism through a protocol application, a collection process

(this would be in agreement with slide 14 (8.EEOO) presented in '2009-02-BiomaterialCoreTerms-Vancouver.ppt')

Comment #1: Extracted seems to imply that a fraction/portion is collected from an organism
If we were to use 'derived entity of organismal origin', we might be able to cover bioengineered organism, should we wish to do. (e.g. we currently have transgenic organism asserted under EOO.the same entity asserted under extracted entity of organism origin would be odd).
Comment #2: Necessary and Sufficient conditions may need to be altered to reflect the changes.

-Organization:
stays where it is for now but feels odd to have it under material entity.
proposition to move if to BFO: possibly as a defined class. 
The current editor notes capture the issues very well.