Another IAO_0000117->IAO_0000116 fix
Multiline value for (definition editor)(!obi:IAO_0000117) on (liquid handler)(!obi:OBI_0400112):
-----
DS: Is this class justified? Its a unnamed class. If so, put a fluidic_system and the fluidic_subsystem as subclasses.
TW: This is required by MO.
FG & DS: Capture as function.
All: Needs to be reviewed, according to query use case. If we keep it its kept as unnamed owl class.
The liquid handling class remains but as an undefined class with are unlikely to have children. It is expected that the reasoner would classifiy appropriate classes under this class that meet the have the liquid_handling function relation.
-----